pyproject.toml
[project]
name = "tellsee-prod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apswutils>=0.1.2",
    "monsterui>=1.0.37",
    "pydantic-ai>=1.44.0",
    "python-dotenv>=1.2.1",
    "python-fasthtml>=0.12.39",
    "supabase>=2.27.2",
    "upstash-redis>=1.5.0",
]
